atc/metric: structure: use int for Counter
On master all the internal metric types use mostly int64's, and we just copied
Counter from master. However, in the era of 5.5.x the internal metric types
still use mostly int's, so let's be consistent.

+++ atc/metric/counter.go
@@ -14,7 +14,7 @@ func (m *Counter) IncDelta(delta int) {
 	atomic.AddInt64((*int64)(&m.cur), int64(delta))
 }
 
-func (m *Counter) Delta() float64 {
+func (m *Counter) Delta() int {
 	cur := atomic.SwapInt64((*int64)(&m.cur), 0)
-	return float64(cur)
+	return int(cur)
 }
